requirements.txt
pypng
opencv-python
opencv-contrib-python
wheel
scikit-learn
tensorflow-gpu==1.14.0
keras==2.2.5
matplotlib
prettytable
ptable
